exports Option and ValueContainer as replaceable components from the DS
