+++ src/Select/Option.scss
@@ -1,0 +1,5 @@
+.Option {
+  display: flex; 
+  justify-content: space-between; 
+  align-items: center;
+}
