+++ src/Select/index.js
@@ -1,11 +1,15 @@
 import AsyncSelect from './AsyncSelect';
 import CreatableSelect from './CreatableSelect';
+import Option from './Option';
 import SingleSelect from './SingleSelect';
+import ValueContainer from './ValueContainer';
 import { SELECT_SIZES } from './styles';
 
 export {
   AsyncSelect,
   CreatableSelect,
+  Option,
   SELECT_SIZES,
   SingleSelect,
+  ValueContainer,
 };
