+++ src/index.js
@@ -28,8 +28,10 @@ import RadioButtonGroup from 'src/RadioButtonGroup';
 import {
   AsyncSelect,
   CreatableSelect,
+  Option,
   SELECT_SIZES,
   SingleSelect,
+  ValueContainer,
 } from 'src/Select';
 import { colors as COLORS } from 'src/Styles';
 import {
@@ -76,6 +78,7 @@ export {
   ModalBody,
   ModalFooter,
   ModalHeader,
+  Option,
   Pill,
   PILL_COLORS,
   Popper,
@@ -94,6 +97,7 @@ export {
   Tooltip,
   TrackedButton,
   useToast,
+  ValueContainer,
   withToast,
   withToastPropTypes,
 };
